+++ CHANGELOG.md
@@ -12,6 +12,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Patch
 
+## [v2.1.0] - 2017-08-13
+
+### Minor
+- Add `catchify.some(iterable)` which is like `catchify.all(iterable)` but 
+an error does not prevent all errors and values from being returned
+
 ## [v2.0.0] - 2017-08-13
 
 ### Major
@@ -68,7 +74,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Major
 - Initial release
 
-[Unreleased]: https://github.com/majgis/catchify/compare/v2.0.0...master
+[Unreleased]: https://github.com/majgis/catchify/compare/v2.1.0...master
+[v2.1.0]: https://github.com/majgis/catchify/compare/v2.0.0...v2.1.0
 [v2.0.0]: https://github.com/majgis/catchify/compare/v1.1.4...v2.0.0
 [v1.1.4]: https://github.com/majgis/catchify/compare/v1.1.3...v1.1.4
 [v1.1.3]: https://github.com/majgis/catchify/compare/v1.1.2...v1.1.3